Catalog now fetch catalog and return CatalogPreview
