+++ src/components/Catalog/__tests__/Catalog.test.js
@@ -1,10 +1,8 @@
 import React from 'react'
-import { shallow, mount } from 'enzyme'
-import Loader from '../../Loader/Loader'
-import CatalogPreview from '../CatalogPreview/CatalogPreview'
+import { mount } from 'enzyme'
+import CatalogPreview from '../CatalogPreview'
 
 import catalog from '../../../fetch/__test__/catalog.json'
-import metrics from '../../../fetch/__test__/metrics.json'
 
 const Catalog = require('proxyquire')('../Catalog', {
   '../../fetch/fetch': require('../../../fetch/__mocks__/fetch')
@@ -13,23 +11,16 @@ const Catalog = require('proxyquire')('../Catalog', {
 describe('<Catalog />', () => {
 
   describe('When all goes well', () => {
-    let wrapper
 
-    beforeEach(() => {
-      wrapper = shallow(<Catalog catalog={catalog} />)
+    it('Should assign catalog to this.state', () => {
+      const wrapper = mount(<Catalog catalogId={'1'} />)
+      const test = <CatalogPreview catalog={catalog} />
+
+      return wrapper.instance()
+        .componentWillMount()
+        .then(() => {
+          expect(wrapper).to.contain(test)
+        })
     })
-
-    it('should display the name of the catalog', () => {
-      expect(wrapper.contains(catalog.name)).to.be.true
-    })
-
-    it('should display <CatalogPreview />', () => {
-      const wrapper = shallow(<Catalog catalog={catalog} />)
-      const catalogPreview = <CatalogPreview metrics={metrics} />
-
-      expect(wrapper.containsMatchingElement(catalogPreview)).to.be.true
-    })
-
   })
-
 })
